Fix Xcode warnings.
TESTED=Run iOS and Mac.

+++ include/screen.h
@@ -34,7 +34,7 @@ class Screen {
   SDL_Renderer *renderer_;
   SDL_Texture *texture_;
   SDL_Window *window_;
-  unsigned int frame_start_ms_ = 0;
+  unsigned long long frame_start_ms_ = 0;
   int frames_ = 0;
   ScreenStyle style_ = ScreenStyle_White;
